servers/easy_journey/requirements.txt
fastapi
uvicorn[standard]
groq
python-dotenv
pydantic
requests
python-multipart
httpx
